♻️ refactor(client/template): 페이지를 구성하는 요소를 템플릿 단위로 분할하고 모든 컴포넌트를 5단계로 분할 시킴
