+++ packages/client/src/App.tsx
@@ -2,7 +2,7 @@ import './App.css';
 import { Route, Router, Routes } from 'react-router-dom';
 import LandingPage from './pages/LandingPage';
 import ProfilePage from './pages/ProfilePage';
-import ChatroomPage from './pages/ChatroomPage';
+import ChatroomPage from './pages/ChatChannelPage';
 
 function App() {
   return (
